+++ src/helpers/processExecutor.ts
@@ -84,14 +84,16 @@ export default class ProcessExecutor {
 
   private static pipeFailed(pid: number | undefined, name: string) {
     return (e: any) => {
-      if (e.code === 'ERR_STREAM_PREMATURE_CLOSE' || e.code === 'EPIPE') {
-        this.logger.trace(
-          'Pipe',
-          name,
-          'of process',
-          pid,
-          'closed prematurely',
-        );
+      if (
+        e.code in
+        [
+          'ERR_STREAM_PREMATURE_CLOSE',
+          'EPIPE',
+          'ERR_STREAM_WRITE_AFTER_END',
+          'ENOENT',
+        ]
+      ) {
+        this.logger.trace(`Pipe ${name} of process ${pid} closed prematurely`);
       } else {
         this.logger.warn('Set up process', pid, name, 'failed', e);
         telemetry.error('pipeFailed', e, { name });